+++ src/components/field/template.jade
@@ -51,41 +51,41 @@ svg(
       +Entity(entity.id, entity.type, entity.x, entity.y, entity.rad)
     +focus(focus.x, focus.y)
     
-  g(
-    transform='translate(' + -cx + ', ' + -cy + ')'
-    key='physics'
-  )
-    for body in bodies
-      - var x = body.pos.x;
-      - var y = body.pos.y;
-      - var med = body.angle;
-      g(
-        key= body.uid
-        transform= 'translate('+x+','+y+') rotate(' + med + ')'
-      )
-        if body.name === "circle"
-          circle(
-            cx=0
-            cy=0
-            r= body.radius
-            fill= 'none',
-            stroke= 'red'
-          )
-        else if body.name === 'rectangle'
-          rect(
-            width=body.width
-            height=body.height
-            x=-body.width/2
-            y=-body.height/2
-            fill='gray'
-          )
-        else if body.name === 'convex-polygon'
-          - var vs = body.vertices;
-          - var points = vs.map(function(v){return v.x+','+v.y;}).join(' ')
-          polygon(
-            points = points
-            fill= 'gray'
-          )
+  //- g(
+  //-   transform='translate(' + -cx + ', ' + -cy + ')'
+  //-   key='physics'
+  //- )
+  //-   for body in bodies
+  //-     - var x = body.pos.x;
+  //-     - var y = body.pos.y;
+  //-     - var med = body.angle;
+  //-     g(
+  //-       key= body.uid
+  //-       transform= 'translate('+x+','+y+') rotate(' + med + ')'
+  //-     )
+  //-       if body.name === "circle"
+  //-         circle(
+  //-           cx=0
+  //-           cy=0
+  //-           r= body.radius
+  //-           fill= 'none',
+  //-           stroke= 'red'
+  //-         )
+  //-       else if body.name === 'rectangle'
+  //-         rect(
+  //-           width=body.width
+  //-           height=body.height
+  //-           x=-body.width/2
+  //-           y=-body.height/2
+  //-           fill='gray'
+  //-         )
+  //-       else if body.name === 'convex-polygon'
+  //-         - var vs = body.vertices;
+  //-         - var points = vs.map(function(v){return v.x+','+v.y;}).join(' ')
+  //-         polygon(
+  //-           points = points
+  //-           fill= 'gray'
+  //-         )
     
 hr
 span score
