+++ src/game/entities/entity.ts
@@ -61,13 +61,4 @@ class Entity extends EventEmitter {
   public dispose(){}
 
   public onHit(other: Entity){}
-
-  public serialize(): {x: number; y:number; rad: number; id: string; type: string;} {
-    return {
-      id: this.id,
-      x: this.x, y:this.y,
-      rad: this.rad,
-      type: (<any>this.constructor).type
-    };
-  }
 }
